+++ editor.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <title>Maskable.app Editor</title>
-    <link href="viewer.css" rel="stylesheet" />
-    <link href="editor.css" rel="stylesheet" />
+    <link href="css/viewer.css" rel="stylesheet" />
+    <link href="css/editor.css" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;900&display=swap" rel="stylesheet" />
     <script type="module" src="src/viewer/libs.js"></script>
     <script type="module" src="src/viewer/change-mask.js"></script>
@@ -15,6 +15,7 @@
     <link rel="preconnect" href="//getinsights.io" />
     <link rel="preload" href="toggle/sun.svg" as="image" />
     <link rel="preload" href="toggle/moon.svg" as="image" />
+    <link rel="modulepreload" href="src/viewer/masks.js" />
     <link rel="modulepreload" href="/web_modules/file-drop-element.js" />
     <link rel="modulepreload" href="/web_modules/dark-mode-toggle.js" />
     <link rel="modulepreload" href="/web_modules/insights-js/dist/esnext/index.js" />
@@ -42,7 +43,7 @@
     <meta name="og:image" content="https://maskable.app/github_social.png" />
     <meta name="og:url" content="https://maskable.app/editor" />
     <meta name="og:description" content="Create maskable icons in the browser!" />
-    <meta name="monetization" content="$pay.stronghold.co/1a123824fc7f4624ae7be3cab5426712a6b">
+    <meta name="monetization" content="$pay.stronghold.co/1a123824fc7f4624ae7be3cab5426712a6b" />
   </head>
 
   <body>
@@ -126,8 +127,8 @@
           </p>
           <p class="about__main">
             <a class="about__link" href="https://css-tricks.com/maskable-icons-android-adaptive-icons-for-your-pwa/">
-              Maskable icons</a
-            >
+              Maskable icons
+            </a>
             allow web developers to specify a full-bleed icon that will be cropped by the user-agent to match other
             icons on the device. On Android, this lets developers get rid of the default white background around their
             icons and use the entire provided space by generating adaptive icons.
