+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>Maskable.app</title>
-    <link href="viewer.css" rel="stylesheet" />
+    <link href="css/viewer.css" rel="stylesheet" />
     <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;900&display=swap" rel="stylesheet" />
     <script type="module" src="src/viewer/libs.js"></script>
     <script type="module" src="src/viewer/upload-icon.js"></script>
@@ -15,6 +15,7 @@
     <link rel="preconnect" href="//getinsights.io" />
     <link rel="preload" href="toggle/sun.svg" as="image" />
     <link rel="preload" href="toggle/moon.svg" as="image" />
+    <link rel="modulepreload" href="src/viewer/masks.js" />
     <link rel="modulepreload" href="/web_modules/file-drop-element.js" />
     <link rel="modulepreload" href="/web_modules/dark-mode-toggle.js" />
     <link rel="modulepreload" href="/web_modules/insights-js/dist/esnext/index.js" />
