+++ masklist.html
@@ -1,0 +1,112 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Change Masks | Maskable.app</title>
+    <link href="css/viewer.css" rel="stylesheet" />
+    <link href="css/masklist.css" rel="stylesheet" />
+    <link href="https://fonts.googleapis.com/css2?family=Lato:wght@400;900&display=swap" rel="stylesheet" />
+    <script type="module" src="src/masklist/previews.js"></script>
+    <script type="module" src="src/viewer/libs.js"></script>
+    <script nomodule src="src/libs-bundle.js" defer></script>
+    <link rel="preconnect" href="//getinsights.io" />
+    <link rel="preload" href="toggle/sun.svg" as="image" />
+    <link rel="preload" href="toggle/moon.svg" as="image" />
+    <link rel="modulepreload" href="/web_modules/dark-mode-toggle.js" />
+    <link rel="modulepreload" href="/web_modules/insights-js/dist/esnext/index.js" />
+
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <link rel="icon" type="image/png" sizes="16x16" href="favicon/favicon_16.png" />
+    <link rel="icon" type="image/png" sizes="32x32" href="favicon/favicon_32.png" />
+    <link rel="icon" type="image/png" sizes="196x196" href="favicon/favicon_196.png" />
+    <link rel="icon" type="image/png" sizes="512x512" href="favicon/favicon_512.png" />
+    <link rel="manifest" href="manifest.json" />
+    <link rel="apple-touch-icon" sizes="196x196" href="favicon/maskable_196.png" />
+    <meta name="theme-color" content="#FFD567" />
+
+    <meta name="monetization" content="$ilp.uphold.com/p2WKq3nJHA8w">
+  </head>
+
+  <body>
+    <main class="viewer">
+      <h1 class="title">Maskable​.app</h1>
+
+      <dark-mode-toggle permanent></dark-mode-toggle>
+
+      <h3>Visible</h3>
+      <ul class="masklist">
+        <li class="mask__drag locked" data-value="none">
+          <div class="mask__preview"></div>
+          None
+        </li>
+        <li class="mask__drag locked" data-value="minimum">
+          <div class="mask__preview"></div>
+          Minimum safe area
+        </li>
+        <li class="mask__drag" data-value="circle">
+          <div class="mask__preview"></div>
+          Circle
+        </li>
+        <li class="mask__drag" data-value="rounded_rect">
+          <div class="mask__preview"></div>
+          Rounded Rectangle
+        </li>
+        <li class="mask__drag" data-value="sharp_rect">
+          <div class="mask__preview"></div>
+          Square
+        </li>
+        <li class="mask__drag" data-value="drop">
+          <div class="mask__preview"></div>
+          Drop
+        </li>
+      </ul>
+
+      <h3>Extra</h3>
+      <ul class="masklist">
+        <li class="mask__drag" data-value="cylinder">
+          <div class="mask__preview"></div>
+          Cylinder
+        </li>
+        <li class="mask__drag" data-value="squircle">
+          <div class="mask__preview"></div>
+          Squircle
+        </li>
+      </ul>
+
+      <section class="controls text">
+        <button type="button" class="button--primary" name="reset">
+          Reset to default
+        </button>
+      </section>
+
+      <hr />
+      <aside class="about text">
+        <p class="about__main">
+          Control which masks will appear inside <a class="about__link" href=".">Maskable.app</a> and the <a class="about__link" href="editor">Editor</a>.
+        </p>
+        <p class="about__extra">
+          Edit on
+          <a class="about__link" href="https://github.com/NotWoods/maskable">
+            GitHub
+          </a>
+        </p>
+        <p class="about__extra">
+          Created by
+          <a class="about__link" href="https://tigeroakes.com">
+            <img class="about__self-logo" src="tigeroakes.svg" alt="Tiger Oakes" height="20" width="135" />
+          </a>
+        </p>
+      </aside>
+    </main>
+
+    <svg height="0" width="0" version="1.1" xmlns="http://www.w3.org/2000/svg">
+      <defs>
+        <clipPath id="squircle">
+          <path
+            d="M96 179.8c36.7 0 57.1 0 70.5-13.3 13.3-13.4 13.3-33.8 13.3-70.5s0-57.1-13.3-70.5C153 12.2 132.7 12.2 96 12.2s-57.1 0-70.5 13.3C12.2 39 12.2 59.3 12.2 96s0 57.1 13.3 70.5C39 179.8 59.3 179.8 96 179.8z"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  </body>
+</html>
