+++ src/viewer/libs.js
@@ -12,7 +12,9 @@ const ad = document.querySelector('[data-ea-publisher]');
  */
 function updateDarkModeClasses(darkMode) {
   document.body.classList.toggle('dark', darkMode);
-  ad.classList.toggle('dark', darkMode)
+  if (ad) {
+    ad.classList.toggle('dark', darkMode);
+  }
 }
 
 // Initialize the toggle based on `prefers-color-scheme`, defaulting to 'light'.
@@ -20,14 +22,14 @@ toggle.mode = matchMedia('(prefers-color-scheme: dark)').matches
   ? 'dark'
   : 'light';
 // Set or remove the `dark` class the first time.
-updateDarkModeClasses(toggle.mode === 'dark')
+updateDarkModeClasses(toggle.mode === 'dark');
 // Listen for toggle changes (which includes `prefers-color-scheme` changes)
 // and toggle the `dark` class accordingly.
 toggle.addEventListener('colorschemechange', () => {
   updateDarkModeClasses(toggle.mode === 'dark');
 });
 
-if (document.monetization) {
+if (document.monetization && ad) {
   document.monetization.addEventListener('monetizationstart', () => {
     if (document.monetization.state === 'started') {
       console.log('Payment started, hiding ads');
